flexibility and customization TestingEvent
Added event delegation functionality to MapNode class

The changes include the addition of a new member variable, "onPlayerEnter", which is of type "MapNodeEvent". This variable is used to store a list of methods that will be called when a player enters a node. The "PlayerEntered()" method is then used to trigger these methods.

Additionally, two new methods, "OnPlayerEnterNode()" and "AnotherMethod()", are added to the class. These methods are then added to the "onPlayerEnter" delegate using the "+=" operator.